add error exceptions when properties do not exist

+++ src/models/properties/properties.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { Prisma } from '@prisma/client';
 import { PrismaService } from 'src/services/prisma/prisma.service';
 
@@ -15,14 +15,43 @@ export class PropertiesService {
   }
 
   async getProperty(id: string) {
-    return this.prisma.property.findUnique({ where: { id } });
+    const property = await this.prisma.property.findUnique({ where: { id } });
+
+    if (!property) {
+      throw new HttpException(
+        'Property with this id does not exist',
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    return property;
   }
 
   async editProperty(id: string, property: Prisma.PropertyUpdateInput) {
+    const existingProperty = await this.prisma.property.findUnique({
+      where: { id },
+    });
+
+    if (!existingProperty) {
+      throw new HttpException(
+        'Property with this id does not exist',
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
     return this.prisma.property.update({ where: { id }, data: property });
   }
 
   async deleteProperty(id: string) {
+    const property = await this.prisma.property.findUnique({ where: { id } });
+
+    if (!property) {
+      throw new HttpException(
+        'Property with this id does not exist',
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
     return this.prisma.property.delete({ where: { id } });
   }
 }
